fix dw_spi_eeprom_read_dma_start type-mismatch

+++ plat/seehi/rhea/drivers/spi/dw_apb_ssi/src/dw_apb_ssi.c
@@ -618,7 +618,7 @@ int16_t dw_spi_tuning(spi_id_t spi_id, void *t_buf, uint8_t t_size, void *r_buf,
 #include "dma.h"
 uint8_t dma_channel_state[8] = {0};
 
-DMA_Channel_t dw_spi_eeprom_read_dma_start(spi_id_t spi_id, void *t_buf, uint8_t t_size, void *r_buf, uint32_t r_size)
+DMA_Channel_t dw_spi_eeprom_read_dma_start(spi_id_t spi_id, void *t_buf, uint8_t t_size, void *r_buf, uint16_t r_size)
 {
 	//0. 判断输入合法性
 	assert(get_spi_state(spi_id)!=0);
